feat: updates code before digging into the bug
