+++ .github/workflows/update.yml
@@ -11,7 +11,7 @@ jobs:
 
     strategy:
       matrix:
-        node-version: [18.x]
+        node-version: [20.x]
 
     steps:
       - uses: actions/checkout@v3
@@ -26,7 +26,7 @@ jobs:
       - run: pnpm install --no-frozen-lockfile
       - run: pnpm build
       - run: pnpm test
-      - run: pnpm update
+      - run: pnpm run update
       - name: Create Pull Request
         uses: peter-evans/create-pull-request@v4
         with:
