+++ package.json
@@ -27,7 +27,7 @@
     "preview": "pnpm --filter @pastoralist/app app-preview",
     "release": "release-it",
     "test": "turbo run test-src",
-    "test-src": "vitest run",
+    "test-src": "jest --coverage --maxWorkers=75%",
     "typecheck": "turbo run typecheck-src",
     "typecheck-src": "tsc --noEmit",
     "update": "codependence --update"
@@ -52,35 +52,37 @@
   },
   "homepage": "https://jeffry.in/pastoralist/",
   "dependencies": {
-    "commander": "^9.5.0",
-    "compare-versions": "^5.0.3",
-    "fast-glob": "^3.2.12",
-    "react-icons": "^4.8.0",
-    "tslib": "^2.5.0"
+    "commander": "^11.0.0",
+    "compare-versions": "^6.0.0",
+    "fast-glob": "^3.3.0"
   },
   "devDependencies": {
-    "@commitlint/cli": "^17.5.1",
-    "@commitlint/config-conventional": "^17.4.4",
-    "@types/node": "^18.15.11",
-    "@typescript-eslint/eslint-plugin": "^5.57.0",
-    "@typescript-eslint/parser": "^5.57.0",
+    "@commitlint/cli": "^17.6.6",
+    "@commitlint/config-conventional": "^17.6.6",
+    "@swc/core": "^1.3.68",
+    "@swc/jest": "^0.2.26",
+    "@types/jest": "^29.5.3",
+    "@types/node": "^20.4.2",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
     "codependence": "^0.2.6",
-    "eslint": "^8.37.0",
+    "eslint": "^8.44.0",
     "eslint-config-prettier": "^8.8.0",
     "husky": "^8.0.3",
     "is-ci": "^3.0.1",
-    "lint-staged": "^13.2.0",
+    "jest": "^29.6.1",
+    "lint-staged": "^13.2.3",
     "path-exists-cli": "^2.0.0",
-    "prettier": "^2.8.7",
-    "release-it": "^15.9.3",
-    "rimraf": "^3.0.2",
+    "prettier": "^3.0.0",
+    "release-it": "^16.1.0",
+    "rimraf": "^5.0.1",
     "stdouttojson": "^0.7.2",
     "ts-node": "^10.9.1",
-    "tsup": "^6.7.0",
-    "turbo": "^1.8.8",
-    "typescript": "^4.9.5",
-    "vite": "^3.2.5",
-    "vitest": "^0.25.8"
+    "tslib": "^2.6.0",
+    "tsup": "^7.1.0",
+    "turbo": "^1.10.7",
+    "typescript": "^5.1.6",
+    "vite": "^4.4.3"
   },
   "commitlint": {
     "extends": [
@@ -92,14 +94,38 @@
       "eslint --fix"
     ]
   },
+  "jest": {
+    "transform": {
+      "^.+\\.(js|ts)$": "@swc/jest"
+    },
+    "testEnvironment": "jest-environment-node",
+    "clearMocks": true,
+    "moduleFileExtensions": [
+      "js",
+      "ts",
+      "json"
+    ],
+    "collectCoverageFrom": [
+      "<rootDir>/src/**/*.ts"
+    ]
+  },
+  "eslintConfig": {
+    "root": true,
+    "parser": "@typescript-eslint/parser",
+    "plugins": ["@typescript-eslint"],
+    "extends": [
+      "eslint:recommended",
+      "plugin:@typescript-eslint/recommended",
+      "prettier"
+    ]
+  },
   "release-it": {
     "git": {
-      "commitMessage": "chore: release v%s"
+      "commitMessage": "chore: release"
     }
   },
   "engines": {
-    "node": ">=16.0.0",
-    "pnpm": ">=7.0.0"
+    "pnpm": ">=8.0.0"
   },
   "resolutions": {
     "react": "^18.2.0",
