+++ page/app/package.json
@@ -12,31 +12,32 @@
     "app-preview": "vite preview"
   },
   "dependencies": {
-    "@mdx-js/react": "^2.1.2",
+    "@mdx-js/react": "^2.3.0",
     "prism-react-renderer": "^1.3.5",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.10.1"
   },
   "devDependencies": {
-    "@mdx-js/mdx": "^2.1.2",
-    "@mdx-js/rollup": "^2.1.2",
-    "@tailwindcss/typography": "^0.5.4",
-    "@types/mdx": "^2.0.2",
-    "@types/node": "^18.6.4",
-    "@types/react": "^18.0.16",
-    "@types/react-dom": "^18.0.6",
-    "@vitejs/plugin-react": "^2.0.0",
-    "autoprefixer": "^10.4.8",
-    "daisyui": "^2.24.0",
-    "mdx-mermaid": "^1.3.2",
-    "mermaid": "^9.1.6",
-    "postcss": "^8.4.16",
-    "react-daisyui": "^2.3.2",
+    "@mdx-js/mdx": "^2.3.0",
+    "@mdx-js/rollup": "^2.3.0",
+    "@tailwindcss/typography": "^0.5.9",
+    "@types/mdx": "^2.0.5",
+    "@types/node": "^20.4.2",
+    "@types/react": "^18.2.14",
+    "@types/react-dom": "^18.2.7",
+    "@vitejs/plugin-react": "^4.0.3",
+    "autoprefixer": "^10.4.14",
+    "daisyui": "^3.2.1",
+    "mdx-mermaid": "^2.0.0",
+    "mermaid": "^10.2.4",
+    "postcss": "^8.4.25",
+    "react-daisyui": "^4.0.0",
     "react-flow-renderer": "^10.3.17",
-    "rehype-slug": "^5.0.1",
+    "rehype-slug": "^5.1.0",
     "remark-gfm": "^3.0.1",
-    "tailwindcss": "^3.1.8",
-    "typescript": "^4.6.4",
-    "vite": "^3.0.0"
+    "tailwindcss": "^3.3.2",
+    "typescript": "^5.1.6",
+    "vite": "^4.4.3"
   }
 }
