+++ page/app/src/components/Drawer.tsx
@@ -1,13 +1,18 @@
 import { Drawer as DaisyDrawer } from "react-daisyui";
 import { SideBar } from "./SideBar";
 
-export const Drawer = ({ children, isOpen, toggleDrawer }: any) => {
+export type DrawerProps = {
+  children: React.ReactNode;
+  isOpen: boolean;
+  toggleDrawer: () => void;
+}
+
+export const Drawer = ({ children, isOpen, toggleDrawer }: DrawerProps) => {
   return (
     <DaisyDrawer
       side={<SideBar />}
       open={isOpen}
       onClickOverlay={toggleDrawer}
-      mobile
     >
       <div className="h-100 px-6">{children}</div>
     </DaisyDrawer>
