+++ src/__tests__/bar-package.json
