+++ src/__tests__/foo-package.json
